[972] Ядро / База: Фикс загрузки SoundOverrideSubclass из item_template ФИкс загрузку таблицы phase_definitions Фикс загрузки guild_xp_for_level и research_loot

+++ src/server/game/Battlegrounds/BattlegroundMgr.cpp
@@ -1015,7 +1015,7 @@ void BattlegroundMgr::CreateInitialBattlegrounds()
         data.MaxPlayersPerTeam = fields[2].GetUInt16();
         data.LevelMin = fields[3].GetUInt8();
         data.LevelMax = fields[4].GetUInt8();
-        uint8 spawn = fields[9].GetUInt8();
+        float spawn = fields[9].GetFloat();
         data.StartMaxDist = float(spawn * spawn);
 
         data.scriptId = sObjectMgr->GetScriptId(fields[11].GetCString());
